# src/docs/swagger.yml
openapi: 3.0.4
info:
  title: Argos-tracker API
  description: |-
    Argos Tracker es un sistema de rastreo y gestion de mascotas diseñado para que cada usuario pueda administrar
    sus collares de rastreo y obtener informacion en tiempo real desde la estacion base.

    La API permite a las aplicaciones cliente (movil o web) interactuar con el sistema de rastreo, gestionar usuarios, dispositivos y sus eventos. 

    Algunas funcionalidades clave: 
      - Registro y autenticacion de usuarios 
      - Gestion de collares asociados a mascotas 
      - Configuracion de collares 
      - Consulta de historial de eventos

    Links utiles:
    - [Repositorio argos-tracker](https://github.com/martinBustos23/argos-tracker)
    - [Documentacion del proyecyo](https://github.com/martinBustos23/argos-tracker)

    Desarrolladores:
    - Martin Bustos: bustiger@argos-tracker.com
    - Maximiliano Ariza: gaspanzer@argos-tracker.com 

    Esta definicion esta basada en el estandar OpenAPI 3.0 y servira como base de documentacion y pruebas en Swagger.
  version: 1.0.0
  contact:
    name: the team
    email: soporte@argos-tracker.com
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /api/users:
    post:
      tags:
        - Users
      summary: Crea un nuevo usuario.
      description: Registra un nuevo usuario en el sistema. Solo administradores pueden crear usuarios.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        description: Datos del nuevo usuario.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              usuario_normal:
                summary: Usuario normal
                value:
                  username: 'Juan'
                  password: 'Constraseña123'
                  admin: false
              usuario_admin:
                summary: Usuario administrador
                value:
                  username: 'admin_nuevo'
                  password: 'admin'
                  admin: true
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos de entrada invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                faltan_parametros:
                  value:
                    error: 'Faltan parametros'
                password_debil:
                  value:
                    error: 'La contraseña no cumple con los requisitos minimos'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Usuario no autorizado'
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Usuario no es administrador'
        '409':
          description: Conflicto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Usuario ya registrado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Users
      summary: Obtiene lista de usuarios
      description: Retorna todos los usuarios activos o inactivos segun el parametro 'active'
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: active
          in: query
          description: Filtrar usuarios por estado 'active' (true = activos, false = inactivos)
          required: false
          schema:
            type: string
            enum: [true, false]
            default: true
          examples:
            activos:
              value: 'true'
              summary: Usuarios activos
            inactivos:
              value: 'false'
              summary: Usuarios inactivos
      responses:
        '200':
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              examples:
                usuarios_activos:
                  summary: Ejemplo de usuarios activos
                  value:
                    - id: 1
                      username: 'admin'
                      image: 'default'
                      admin: true
                      active: true
                      lastLogin: '2025-10-04T12:55:40.000Z'
                      timeout: null
                    - id: 2
                      username: 'Juan'
                      image: 'default'
                      admin: false
                      active: true
                      lastLogin: '2025-10-03T18:30:15.000Z'
                      timeout: null
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Usuario no autorizado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint de login
  schemas:
    UserCreate:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 32
          description: Nombre de usuario
          example: 'Juan'
        password:
          type: string
          minLength: 8
          maxLength: 256
          description: Contraseña del usuario
          example: 'Contraseña123'
        image:
          type: string
          description: Nombre de la imagen del usuario para la URL (default si no se especifica)
          default: 'default'
          example: 'default'
        # admin:
        #   type: boolean
        #   description: Define si el usuario tiene rol de administrador
        #   example: false
      description: Es probable que mas parametros sean requeridos para la versiona final de la app... por ejemplo admin
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: Identificador unico del usuario
          example: 2
        username:
          type: string
          description: Nombre de usuario
          example: 'Juan'
        image:
          type: string
          description: Nombre de la imagen del usuario para la URL
          example: 'default'
        admin:
          type: boolean
          description: Indica si el usuario es administrador
          example: false
        active:
          type: boolean
          description: Indica si la cuenta esta activa
          example: true
        lastLogin:
          type: string
          format: date-time
          nullable: true
          description: Fecha del ultimo inicio de sesión
          example: '2025-10-04T12:55:40.000Z'
        timeout:
          type: string
          format: date-time
          nullable: true
          description: Fecha hasta cuando el usuario esta bloqueado temporalmente
          example: null
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo
          example: 'Error interno del servidor'